fix: improve readability post model
